Fix rendering workflow after cold boot (#2159)
This PR fixes #1940. The cause of a workflow not being able to render
after a cold boot is the missing operator metadata when rendering
starts. At the start of `OperatorMetadataService`, it requests the
backend for operator metadata. And the backend
`/resources/operator-metadata` requires authentication. However, when
OperatorMetadataService is initialized, there might not be any user in
`UserService` yet. Thus, the initialization HTTP call may return 401,
and no metadata is received.

This PR removes the authentication on `/resources/operator-metadata`.
This API can be invoked with or without user information.