User Quota Dashboard (#2193)
### Introduction
---
This PR introduces Quota Dashboard for both users and administrators to
visualize storage usage in Texera. This dashboard allows the users to:

- View their quota (# of created workflows, # of uploaded files, ...)
- View the storage size of the operator cache in their workflows.
- View the Top Five Files that they have uploaded, sorted by the size of
each file.
- Delete Operator Cache.
- Delete and download the files.

The users can only see their own quota dasboard, whereas the
administrators are able to see the quota dashboards of all users.

### Solution
---
The main goal of this dashboard is to have the users to visualize their
storage usage, and organize their storage. There are three tabs in the
dashboard, which are Quota Table, Operator Cache, and Top 5 Files. Each
tab has different strucutre and shows different information.

- `Quota Table`: Displaying boxes with different information in each box
- `Files Uploaded`: Displaying the number of the files the user
uploaded.
- `Workflows Created`: Displaying the number of workflows the user
created.
- `Files with Access`: Displaying the number of files the user has
access to.
- `Workflows with Access`: Displaying the number of workflows the user
has access to.
- `Total Size of the Files`: Displaying the total size of the uploaded
files.
- `Total Operator Cache Size`: Displaying the total size of the operator
cache in MongoDB.
- `Operator Cache`: DIsplaying in a hierarchical structure: each
workflow has a table of executions
- `Collection Name`: Name of the collection (operator cache) in MongoDB.
  - `Execution ID`: Different ID for each executions of a workflow.
  - `Cache Size`: The corresponding size of each collection in MongoDB.
  - `Action`: Give the user option to delete collection from MongoDB.
- `Top 5 Files`: Displaying a table consisting 5 largest files.
  - `File Name`: Displaying the name of each file.
  - `Size`: Displaying the size of each file.
  - `Uploaded Time`: The date and time the file uploaded by the user.
- `File Description`: Displaying the description of the files.
Displaying message "No Available Description" when no descriptions
provided.
- `Action`: Provide the user options to either delete or download the
file.

When user clicks the deletion or download button, there will be a pop up
confirmation window.

In addition to the Quota Dashboard Intergace, this PR also modified the
backend clean up code. After starting new execution, the older
executions' results would not be deleted. The clean up of the old
results in mongodb are left to the periodic clean up.

Current PR contains the backend codes, and the frontend codes are in
this PR #2194

### Demo
---



https://github.com/Texera/texera/assets/86388854/eb61acf8-16ad-4a29-880f-2b6575ca17c8

---------

Co-authored-by: MiuMiuMiue <yuanzhe2001@gmail.com>