Expose port to LinkIdentity to allow connection on a port of an internal layer (#2156)
A logical operator may have multiple layers, its physical plan can be
viewed as a sub-DAG. Previously, when we connected the physical plans of
each logical operator, we always assumed that the link would be
connected from the previous sub-DAG's terminal operator output port 0 to
the next sub-DAG's first operator input port 0.

This assumption may not be held in the future as we need to have a link
that connects directly to an internal layer's port arbitrarily. This PR
relaxes this assumption by refactoring the `LinkIdentity(fromLayer,
toLayer)` to `LinkIdentity(fromLayer, fromPort, toLayer, toPort)`. In
this way, the links can be used to connect to internal ports as well.

A use case is to support two-layer `IFOperator`, which has a Python
Reducer layer (physical operator) to evaluate the condition and an `IF`
layer (physical operator) to route the data tuples. In this case, the
data tuples (coming from S1) need to connect directly to the `IF` layer.


![mutilayerif](https://github.com/Texera/texera/assets/11544314/800546a5-35f8-4549-a388-088fd209d5d0)

We now distinguish two types of links, and users need to specify them in
OpDesc:
1. **inter-operator links**: Those links are specified by the user in
DESC. They are described as `OperatorLink`.
2. **intra-operator links**: Those links connect layers of a logical
operator. They are described in `operatorExecutorMultiLayer`.

This PR is the first step of the refactoring. There are some assumptions
to be relaxed later:
1. In this PR, we rely on using `portName` when connecting to internal
ports. This is because we currently do not have a mapping relationship
between the logical operator's ports and the internal layers' ports.
2. The pipelined regions are still using port 0 (for both input/output)
as the default port to add links. This assumption currently holds
because all blocking edge use cases have 1 input port only
(MaterializationRewriter, etc). All operators only have 1 output port.
3. The internal ports' dependencies are interitating external ports'
dependencies. This may need to be relaxed soon.

---------

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>
Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>
Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>
Co-authored-by: Shengquan Ni <shengqun@uci.edu>