Merge DebuggerFrame into ConsoleFrame (#2247)
This PR merges debugger features from `DebuggerFrame` into
`ConsoleFrame` to enhance user experience. The user will be able to
conduct all debugging related functionalities right in the
`ConsoleFrame`, which will give the user more context when interacting
with the debugger.

Now we move the two buttons: `Skip` and `Retry` next to the console
input box. Currently, those two buttons will always be shown no matter
the state of the operator. As `Skip`'s backend logic is disabled now,
the button is disabled for now.

![CleanShot 2023-12-03 at 19 56
15](https://github.com/Texera/texera/assets/17627829/5fe01899-c4d0-4fbd-a98b-d3e8a2e42f78)

**The other features in `DebuggerFrame` including
`PythonExpressionEvaluate`, and the frame itself now have been
deleted.**