Hot-reload configurations in application.conf  (#2245)
## Hot-reloading of configurations

This PR refactors the old Constants global object where we store all the
configuration as static variables. Now all configurations are defined as
methods and will trigger a file modification check. If the
`application.conf` is modified, the updated file will be parsed and new
values of configurations will be retrieved.

### Refactoring of Constants
This PR removes the old Constants object and creates a new global object
called AmberConfig. All configurations are retrieved by this object. The
developer is not allowed to retrieve the values directly from
`application.conf` with the string path such as
`flow-control.max-credit-allowed-in-bytes-per-channel`. This enforces
all configurations are retrieved with the hot-reload mechanism.

### Lazy reloading of configuration file
The AmberConfig object maintains the last modified timestamp of
`application.conf` and checks if the timestamp remains the same each
time a configuration is retrieved. If not, it re-loads the
configuration, otherwise, it retrieves the value from a cached copy.

**IMPORTANT NOTE: For safety, this hot-reload feature now only works on
the following configurations:**
- flow-control.max-credit-allowed-in-bytes-per-channel
- network-buffering.default-batch-size

For other configurations, we can revisit hot-reload when needed.