Refactor workflow compiler with clean data classes (#2255)
This PR refactors the WorkflowCompiler into clear data classes,
including input and outputs:

### Compiler Input
1. **LogicalPlanPojo**, containing information about the operators,
links, breakpoints, and reuse and cache operator ids. The information is
not a complete DAG yet.
2. **WorkflowContext**, containing information regarding the workflow,
including user id, execution id, etc.

### Compiler Output
1. **Logical Plan**, a valid DAG with sources, sinks, other operators,
links, and schemas of the operators. It contains the following mappings:
    - `Map[String, OperatorDescriptor]` where String is the operator Id;
- `DirectedAcyclicGraph[String, OperatorLink]` where the String is the
operator Id.
    
_Note, there are two Logical Plans outputed, one before cache rewritting
and one after cache rewritting. This is for the purpose of comparing
consecutive executions and determine which operator can be replaced by
cache. To be relaxed and refactored later._

3. **Physical Plan**, a valid DAG with layers of operators
(OpExecConfig) expanded, physical links (LinkIdentity). It contains the
mappings:
    - `Map[LayerIdentity, OpExecConfig]`;
    - `DirectedAcyclicGraph[LayerIdentity, DefaultEdge]`.
4. **ExecutionPlan**, a valid DAG of Regions. It contains the mapping:
`DirectedAcyclicGraph[PipelinedRegion, DefaultEdge]`.
5. **PartitioningPlan**, a mapping for each link strategies for worker
deployment: `Map[LinkIdentity, LinkStrategy]`. (This is tentative, to be
replaced by resource allocator later).

### Minor fix
During the refactoring, this PR also fixes an issue that breaks the
workflow cache. The cause is the schema in the logical plan is not being
updated properly.