Introduce RegionConfig and WorkerConfig (#2270)
Previously, for an execution, its configuration is set on each physical
operator. For example, the developer gives the number of the workers of
a physical operator when constructing it. This PR aims to extract the
configurations from the physical operator so that components like
ResourceAllocator can dynamically decide the configurations for an
execution.

### New config data classes
1. `RegionConfig`. A RegionConfig consists of all the configurations of
the physical operators within the region (as well as those effective
operators, e.g., downstream operators affected by blocking links). **We
use the length of the workersConfigs list as the number of workers**.
```
case class RegionConfig(
    workerConfigs: Map[PhysicalOpIdentity, List[WorkerConfig]]
)
```
2. `WorkerConfig`. A WorkerConfig consists of all the configurations of
the worker. In this PR we only migrated the existing configs from
`WorkflowWorkerConfig`. We can add more configurations such as batch
size and flow control queue size here in future PRs.

### Life cycle of assigning `numWorker`.
Now we remove the `numWorker` setting by the developer on a physical
operator. Each physical operator has a field `parallelizable: Boolean`
indicating whether it can be run in parallel. The actual number of
workers is determined by the RegionGenerator (internally, it should be
the ResourceAllocator in the future). The life cycle is as followed:
1. Developer marks if a physical operator is parallelizable.
2. A region generator will generate regions with configs. The physical
operators are assigned with workerIds corresponding to the config.
3. When building a region, the actual worker is created according to the
assigned workerIds.

**suggestedWorkerNum**: To be backward compatible with the worker number
assigned from the UI (in Python UDFs), we allow PhysicalOp to have this
optional field to give a hint. For now, to keep the same behavior, the
generated RegionConfig will respect this suggested worker number and
generate the exact number of workers as suggested. This might be relaxed
or changed in the future (i.e., it might become a hint that the
generated config will consider but may not respect it, or we might
remove it in the future).