Display Runtime Statistics of Each Workflow Execution in a Line Chart Format (#2273)
This PR introduces a new feature that displays runtime statistics of
each workflow execution in the form of a line chart. To achieve this, a
new column called "Runtime Statistics" has been added to the execution
results table.

When a user clicks the line chart icon, the frontend sends an HTTP
request to the backend with the workflow ID and execution ID. The
backend then queries the MySQL database to retrieve the runtime
statistics of the particular execution and sends them back to the
frontend. The frontend uses these statistics to generate a line chart
using the "Plotly.js" library.

The x-axis of the line chart represents the time interval. Currently,
the interval is set to 0.5 seconds, but this can be changed to 2 seconds
in the future. The y-axis represents the metric of the statistics, which
can be switched by clicking on the tab.

It is important to note that the values of the input tuple count and
output tuple count have been aggregated for each time interval. This
means that the line chart shows the total number of input and output
tuple counts processed by each operator.


![미디어1](https://github.com/Texera/texera/assets/143021053/13e4cabf-8987-4cd0-ab9c-c4c5a89fd09f)