Introduce LinkConfig and ChannelConfig (#2282)
This PR adds `LinkConfig` as part of the `RegionConfig`. Previously, a
PhysicalLink's properties (such as partitioning) are stored directly in
the Link itself. As we delay the number of workers into a later stage
(ResourceAllocator) to be dynamically decided, the partitioning of data
between those workers would also be decided later in ResourceAllocator.
Thus we extracted the `LinkConfig` to store the information that is
dynamically generated for a link.

A LinkConfig describes the channels and their partitioning requirements.
```
case class LinkConfig(
    channelConfigs: List[ChannelConfig],
    partitioning: Partitioning
)
```

A ChannelConfig describes a (data) channel's properties between two
workers.
```
case class ChannelConfig(
    fromWorkerId: ActorVirtualIdentity,
    toWorkerId: ActorVirtualIdentity
)
```

`ResourceAllocator` is responsible for generating all the `LinkConfig`
in a region together with the `WorkerConfig`, by considering the
resources taken by workers and channels.

As we require workerIds in ChannelConfig, we are now adding `workerId`
into `WorkerConfig` as well. They are one-to-one matched: one workerId
corresponds to exactly one `WorkerConfig`:
```
case class WorkerConfig(
    workerId: ActorVirtualIdentity
)
```

## Updated DefaultResourceAllocator Logic:
The default resource allocator
1. First decides `List[WorkerConfig]` for each physical operator of a
region.
2. Then it propagates the partition requirements on links in a
topological manner for a region. Incrementally, it will eventually
propagate the entire physical plan.
3. The derived partitioning requirements are later translated into
corresponding `LinkConfig(List[ChannelConfig], Partitioning)`.