Introduce OperatorConfig (#2296)
This PR introduces `OperatorConfig`, as part of the `RegionConfig`. The
previous `List[WorkerConfig]` is now moved inside the corresponding
`OperatorConfig`.

## OperatorConfig and RegionConfig
An OperatorConfig describes the resources for an operator. Currently we
only have the list of WorkerConfigs, more to be added.
```
case class OperatorConfig(
    workerConfigs: List[WorkerConfig]
)
```

The updated RegionConfig:
```
case class RegionConfig(
    operatorConfigs: Map[PhysicalOpIdentity, OperatorConfig],
    linkConfigs: Map[PhysicalLinkIdentity, LinkConfig]
)
```


## New operator initialization life cycle
OperatorConfig is generated by the ResourceAllocator, and to be used at
a few places:
1. `OpExecInitInfo` now takes in an `OperatorConfig` besides the index
and the physicalOp.
2. `WorkflowWorker.prop` now takes in an `OperatorConfig`.

## Other affected changes
With this change, we now removed workerIds from physicalOp. All the
actorIds are to be found in the configs.